Significant Refactoring around storage of Collections
Removed the option to store a Collection within the application’s Container. This was just a bad idea.

Removing the Master Collection concept, since sandboxing won’t allow the app to arbitrarily open any collection anywhere.

Replacing the Master Collection with the implementation of a Parent Realm. This is a “collection of collections,” as was the Master Collection, but with these differences:

* There is not necessarily a single Parent Realm;
* The Parent Realm is specified as a parent folder, and all collections found within that folder become part of the Realm;
* The Parent Realm is built dynamically, on the user’s request, and is not saved to disk.

Also fixed a few bugs.