Edit Tab UI Tweaks, including the addition of scrolling
Replaced the Note Edit NSView with a Scroll View containing an NSview, in order to allow scrolling of the entire edit view (to ensure all fields are accessible, no matter what size the screen or the window, and no matter how many fields a Collection has). Getting the whole thing to work was tricky. Had to change layout from Translates Mask into Constraints to just plain Automatic. Then had to tweak the NSView bottom constraint to make the Superview.Bottom Less Than or Equal to the Parent View.Bottom.